+++ pkg/server/graph/map.go
@@ -1,6 +1,8 @@
 package graph
 
 import (
+	"fmt"
+
 	"github.com/jjzcru/elk/pkg/primitives/ox"
 	"github.com/jjzcru/elk/pkg/server/graph/model"
 )
@@ -22,7 +24,7 @@ func mapElk(elk *ox.Elk) (*model.Elk, error) {
 	}
 
 	for k, v := range elk.Tasks {
-		task, err := mapTask(v)
+		task, err := mapTask(v, k)
 		if err != nil {
 			return nil, err
 		}
@@ -38,9 +40,10 @@ func mapElk(elk *ox.Elk) (*model.Elk, error) {
 	return &elkModel, nil
 }
 
-func mapTask(task ox.Task) (*model.Task, error) {
+func mapTask(task ox.Task, name string) (*model.Task, error) {
 	taskModel := model.Task{
 		Title:       task.Title,
+		Name:        name,
 		Tags:        uniqueString(task.Tags),
 		Cmds:        []*string{},
 		Env:         map[string]interface{}{},
@@ -97,4 +100,179 @@ func mapDep(dep ox.Dep) *model.Dep {
 	}
 
 	return &depModel
+}
+
+func mapTaskInput(task model.TaskInput) ox.Task {
+	env := make(map[string]string)
+	vars := make(map[string]string)
+
+	var deps []ox.Dep
+	var log ox.Log
+
+	title := ""
+	envFile := ""
+	description := ""
+	dir := ""
+	sources := ""
+
+	ignoreError := false
+
+	if task.Env != nil {
+		for k, v := range task.Env {
+			env[k] = fmt.Sprintf("%v", v)
+		}
+	}
+
+	if task.Vars != nil {
+		for k, v := range task.Vars {
+			vars[k] = fmt.Sprintf("%v", v)
+		}
+	}
+
+	if task.Title != nil {
+		title = *task.Title
+	}
+
+	if task.EnvFile != nil {
+		envFile = *task.EnvFile
+	}
+
+	if task.Description != nil {
+		description = *task.Description
+	}
+
+	if task.Dir != nil {
+		dir = *task.Description
+	}
+
+	if task.Sources != nil {
+		sources = *task.Sources
+	}
+
+	if task.IgnoreError != nil {
+		ignoreError = *task.IgnoreError
+	}
+
+	if task.Deps != nil {
+		for _, dep := range task.Deps {
+			deps = append(deps, ox.Dep{
+				Name:        dep.Name,
+				Detached:    dep.Detached,
+				IgnoreError: dep.IgnoreError,
+			})
+		}
+	}
+
+	if task.Log != nil {
+		logFormat := ""
+
+		if task.Log.Format != nil {
+			logFormat = task.Log.Format.String()
+		}
+
+		log = ox.Log{
+			Out:    task.Log.Out,
+			Err:    task.Log.Error,
+			Format: logFormat,
+		}
+	}
+
+	return ox.Task{
+		Title:       title,
+		Tags:        task.Tags,
+		Cmds:        task.Cmds,
+		Env:         env,
+		Vars:        vars,
+		EnvFile:     envFile,
+		Description: description,
+		Dir:         dir,
+		Sources:     sources,
+		IgnoreError: ignoreError,
+		Log:         log,
+		Deps:        deps,
+	}
+}
+
+func mergeTaskInput(taskInput model.TaskInput, task ox.Task) ox.Task {
+	if taskInput.Title != nil {
+		task.Title = *taskInput.Title
+	}
+
+	if taskInput.Tags != nil {
+		task.Tags = taskInput.Tags
+	}
+
+	if taskInput.Cmds != nil {
+		task.Cmds = taskInput.Cmds
+	}
+
+	if taskInput.Env != nil {
+		env := make(map[string]string)
+		for k, v := range taskInput.Env {
+			env[k] = fmt.Sprintf("%v", v)
+		}
+		task.Env = env
+	}
+
+	if taskInput.Vars != nil {
+		vars := make(map[string]string)
+		for k, v := range taskInput.Vars {
+			vars[k] = fmt.Sprintf("%v", v)
+		}
+		task.Vars = vars
+	}
+
+	if taskInput.EnvFile != nil {
+		task.EnvFile = *taskInput.EnvFile
+	}
+
+	if taskInput.Description != nil {
+		task.Description = *taskInput.Description
+	}
+
+	if taskInput.Dir != nil {
+		task.Dir = *taskInput.Dir
+	}
+
+	if taskInput.Log != nil {
+		var log ox.Log
+
+		logFormat := ""
+
+		if taskInput.Log.Format != nil {
+			logFormat = taskInput.Log.Format.String()
+		}
+
+		log = ox.Log{
+			Out:    taskInput.Log.Out,
+			Err:    taskInput.Log.Error,
+			Format: logFormat,
+		}
+
+		task.Log = log
+	}
+
+	if taskInput.Sources != nil {
+		task.Sources = *taskInput.Sources
+	}
+
+	if taskInput.Deps != nil {
+		var deps []ox.Dep
+
+		for _, dep := range taskInput.Deps {
+			deps = append(deps, ox.Dep{
+				Name:        dep.Name,
+				Detached:    dep.Detached,
+				IgnoreError: dep.IgnoreError,
+			})
+		}
+
+		task.Deps = deps
+	}
+
+	if taskInput.IgnoreError != nil {
+		task.IgnoreError = *taskInput.IgnoreError
+	}
+
+	return task
 }